toggle completed state for each chapter
